# api.yaml
openapi: 3.0.0
info:
  title: Instihire Placement Portal API
  description: API documentation for the Instihire backend (Flask + SQLAlchemy).
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local Development Server
components:
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        is_active:
          type: boolean
...
    Drive:
      type: object
      properties:
        id:
          type: integer
        job_title:
          type: string
        job_description:
          type: string
        company_name:
          type: string
        required_degree:
          type: string
        required_branch:
          type: string
        min_cgpa:
          type: number
        deadline:
          type: string
          format: date-time
        interview_mode:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
security:
  - tokenAuth: []
paths:
  /api/auth/login:
    post:
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /api/auth/register/company:
    post:
      summary: Register as a company
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - company_name
                - hr_contact
                - hr_phone
              properties:
                username:
                  type: string
                password:
                  type: string
                company_name:
                  type: string
                hr_contact:
                  type: string
                hr_phone:
                  type: string
                website:
                  type: string
      responses:
        '201':
          description: Company registered successfully
        '400':
          description: Missing required fields or invalid data
  /api/auth/register/student:
    post:
      summary: Register as a student
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - full_name
                - college_name
                - degree
                - branch
                - cgpa
                - graduation_year
              properties:
                username:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
                college_name:
                  type: string
                degree:
                  type: string
                branch:
                  type: string
                cgpa:
                  type: number
                graduation_year:
                  type: integer
      responses:
        '201':
          description: Student registered successfully
        '400':
          description: Missing/invalid fields
...
  /api/company/drives:
    get:
      summary: List company's own drives
      responses:
        '200':
          description: List of drives created by the company
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drive'
    post:
      summary: Create a new placement drive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - job_title
                - job_description
                - deadline
              properties:
                job_title:
                  type: string
                job_description:
                  type: string
                deadline:
                  type: string
                  format: date-time
                min_cgpa:
                  type: number
                required_degree:
                  type: string
                required_branch:
                  type: string
                interview_mode:
                  type: string
      responses:
        '201':
          description: Drive created successfully
  /api/company/drives/{drive_id}/deadline:
    put:
      summary: Update drive deadline
      parameters:
        - in: path
          name: drive_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                deadline:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Deadline updated successfully
  /api/company/drives/{drive_id}:
    delete:
      summary: Delete a company drive
      parameters:
        - in: path
          name: drive_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Drive deleted successfully
  /api/company/drives/{drive_id}/applications:
    get:
      summary: View applications for a specific drive
      parameters:
        - in: path
          name: drive_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of applicants
  /api/company/applications/{app_id}/status:
    post:
      summary: Update application status
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Shortlisted, Selected, Rejected]
      responses:
        '200':
          description: Status updated successfully
  /api/company/applications/{app_id}:
    delete:
      summary: Delete an application
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application removed successfully
...